+++ tests/stdlib/tsetutils.nim
@@ -17,24 +17,18 @@ template main =
 
   block: # fullSet
     doAssert fullSet(Colors) == {red, green, blue}
+    doAssert fullSet(bool) == {true, false}
     doAssert fullSet(char) == {0.chr..255.chr}
     doAssert fullSet(Bar) == {bar0, bar1, bar2}
 
-  block: # not
-    doAssert {red, blue}.not == {green}
-    doAssert (not {red, green, blue}).card == 0
-    doAssert {bar0}.not == {bar1, bar2}
-    doAssert {range[0..10](0), 1, 2, 3}.not == {range[0..10](4), 5, 6, 7, 8, 9, 10}
-    doAssert {'0'..'9'}.not == {0.char..255.char} - {'0'..'9'}
+  block: # complement
+    doAssert {red, blue}.complement == {green}
+    doAssert (complement {red, green, blue}).card == 0
+    doAssert (complement {false}) == {true}
+    doAssert {bar0}.complement == {bar1, bar2}
+    doAssert {range[0..10](0), 1, 2, 3}.complement == {range[0..10](4), 5, 6, 7, 8, 9, 10}
+    doAssert {'0'..'9'}.complement == {0.char..255.char} - {'0'..'9'}
 
-  block: # bugs
-    template impl =
-      doAssert (not {false}) == {true}
-      doAssert fullSet(bool) == {true, false}
-    when defined(js):
-      when nimvm: impl()
-      else: discard # pending bug #17076
-    else: impl()
 
 main()
 static: main()