+++ changelog.md
@@ -78,7 +78,7 @@
 - Added `setutils.toSet` that can take any iterable and convert it to a built-in `set`,
   if the iterable yields a built-in settable type.
 - Added `setutils.fullSet` which returns a full built-in `set` for a valid type.
-- Added `setutils.not` which returns the complement of a built-in `set`.
+- Added `setutils.complement` which returns the complement of a built-in `set`.
 
 
 - Added `math.isNaN`.